moving selector validation into another file
